+++ lib/zammad_api/resources/ticket_article_attachment.rb
@@ -17,7 +17,7 @@ module ZammadAPI
       # @param attributes [Hash]
       def initialize(transport, attributes = {})
         @transport  = transport
-        @attributes = deep_symbolize(attributes || {})
+        @attributes = frozen_attributes(attributes || {})
       end
 
       # Downloads the attachment.